MenuBarSlider
Added a slider that slides when user selects an item inside the menubar. 

This has been achieved by adding a view that matches the size of a cell and a variable of type NSLayoutConstraint for the leading constraint. The x value of the leading constraint gets multiplied by the cellWidth times the indexPath.item of the selected cell. This gets animated by accessing the layoutIfNeeded property inside the animation.