RetrievedImageView + UIDatePicker
RetrievedImageView, a subclass of UIImageView has been created to hold on to several properties which are shared between the UIImageView used for the images retrieved from the APIs. This reduces code in the viewControllers. (DRY)

The minimum and maximum dates for UIDatePicker have been set to the curiosity-rover landing date and current date respectively.